UserAuth and User models
The user authorization is for backend system is finished. waiting to upload and test with templates.

The user models are half way complete but waiting to see what all we want our users to have to fill in when creating a profile. as of now the basic are implemented and ready to go.